fetchAllCategories hooked up as an action & reducer
