+++ React-Frontend/src/actions/types.js
@@ -1,5 +1,7 @@
 export const FETCH_CATEGORIES = 'FETCH_CATEGORIES';
 export const FETCH_CATEGORY = 'FETCH_CATEGORY';
+export const NEW_FETCH_CATEGORIES = 'NEW_FETCH_CATEGORIES';
+export const SELECT_CATEGORY = 'SELECT_CATEGORY';
 export const FETCH_ASSESSMENTS = 'FETCH_ASSESSMENTS';
 export const FETCH_RECOMMENDATIONS = 'FETCH_RECOMMENDATIONS';
 export const FETCH_SCORES = 'FETCH_SCORES';
