+++ React-Frontend/src/reducers/reducer_categories.js
@@ -1,6 +1,13 @@
-import { FETCH_CATEGORY, FETCH_CATEGORIES } from '../actions/types';
+import {
+	FETCH_CATEGORY,
+	FETCH_CATEGORIES,
+	NEW_FETCH_CATEGORIES,
+	SELECT_CATEGORY
+} from '../actions/types';
 
-const INITIAL_STATE = { titles: null, questions: null, checked: false};
+// After we switch to NEW_FETCH_CATEGORIES, won't need titles or questions.
+// Still need to implement something in here to track 'checked' for each individual category
+const INITIAL_STATE = { titles: '', questions: '', categories: '', selected: '' };
 
 export default function(state = INITIAL_STATE, action) {
 	switch(action.type) {
@@ -8,8 +15,10 @@ export default function(state = INITIAL_STATE, action) {
 			return { ...state, questions: action.payload.data.rows };
 		case FETCH_CATEGORIES:
 			return { ...state, titles: action.payload.data.rows };
-		// case UPDATE_CHECKED:
-		// 	return { ...state, checked: action.payload.checked};
+		case NEW_FETCH_CATEGORIES:
+			return { ...state, categories: action.payload.data};
+		case SELECT_CATEGORY:
+			return { ...state, selected: action.payload};
 		default:
 			return state;
 	}
